+++ sdk/cosmos/azure-cosmos/samples/feed_range_management_async.py
@@ -13,6 +13,9 @@ from azure.cosmos.partition_key import PartitionKey
 
 import config
 
+# Use unique suffixes so the sample never collides with pre-existing resources
+_SUFFIX = str(uuid.uuid4())[:8]
+
 # ----------------------------------------------------------------------------------------------------------
 # Prerequisites -
 #
@@ -48,8 +51,8 @@ import config
 
 HOST = config.settings['host']
 MASTER_KEY = config.settings['master_key']
-DATABASE_ID = config.settings['database_id']
-CONTAINER_ID = config.settings['container_id']
+DATABASE_ID = config.settings['database_id'] + '-feed-range-sample-' + _SUFFIX
+CONTAINER_ID = config.settings['container_id'] + '-feed-range-sample-' + _SUFFIX
 
 # Partition key values used throughout the sample
 PARTITION_KEY_VALUES = ['Seattle', 'Portland', 'Denver', 'Austin', 'Chicago']
@@ -326,23 +329,19 @@ async def parallel_change_feed_processing(container):
 
 async def run_sample():
     async with CosmosClient(HOST, {'masterKey': MASTER_KEY}) as client:
-        try:
-            # Setup database and container
-            try:
-                db = await client.create_database(id=DATABASE_ID)
-            except exceptions.CosmosResourceExistsError:
-                db = client.get_database_client(DATABASE_ID)
+        db = None
 
-            try:
-                container = await db.create_container(
-                    id=CONTAINER_ID,
-                    partition_key=PartitionKey(path='/city'),
-                    offer_throughput=400
-                )
-                print('Container with id \'{0}\' created'.format(CONTAINER_ID))
-            except exceptions.CosmosResourceExistsError:
-                container = db.get_container_client(CONTAINER_ID)
-                print('Container with id \'{0}\' already exists'.format(CONTAINER_ID))
+        try:
+            # Setup database and container (unique IDs so we never touch pre-existing resources)
+            db = await client.create_database(id=DATABASE_ID)
+            print('Database with id \'{0}\' created'.format(DATABASE_ID))
+
+            container = await db.create_container(
+                id=CONTAINER_ID,
+                partition_key=PartitionKey(path='/city'),
+                offer_throughput=400
+            )
+            print('Container with id \'{0}\' created'.format(CONTAINER_ID))
 
             # Create sample data
             await create_sample_items(container)
@@ -368,16 +367,17 @@ async def run_sample():
             # 7. Parallel change feed processing using feed ranges
             await parallel_change_feed_processing(container)
 
-            # Cleanup
-            try:
-                await client.delete_database(db)
-            except exceptions.CosmosResourceNotFoundError:
-                pass
-
         except exceptions.CosmosHttpResponseError as e:
             print('\nrun_sample has caught an error. {0}'.format(e.message))
 
         finally:
+            # Clean up the sample database if it was created by this run
+            if db is not None:
+                try:
+                    await client.delete_database(db)
+                    print('\nSample database \'{0}\' deleted'.format(DATABASE_ID))
+                except exceptions.CosmosResourceNotFoundError:
+                    pass
             print("\nrun_sample done")
